feat: Level06 암호화 구현(User CRUD, Login) #08

+++ src/main/java/com/example/scheduleappserverjpa/service/UserServiceImpl.java
@@ -1,7 +1,9 @@
 package com.example.scheduleappserverjpa.service;
 
+import com.example.scheduleappserverjpa.config.PasswordEncoder;
 import com.example.scheduleappserverjpa.dto.user.*;
 import com.example.scheduleappserverjpa.entity.User;
+import com.example.scheduleappserverjpa.exception.InvalidPasswordException;
 import com.example.scheduleappserverjpa.exception.InvalidRequestException;
 import com.example.scheduleappserverjpa.repository.UserRepository;
 import io.micrometer.common.util.StringUtils;
@@ -20,10 +22,15 @@ import java.util.Optional;
 public class UserServiceImpl implements UserService {
 
   private final UserRepository userRepository;
+  private final PasswordEncoder passwordEncoder;
 
   @Override
   public SignUpResponseDto signUp(SignUpRequestDto dto) {
-    User user = new User(dto.getName(), dto.getEmail(), dto.getPwd());
+    // pwd 암호화
+    String encode = passwordEncoder.encode(dto.getPwd());
+    // user Entity 생성
+    User user = new User(dto.getName(), dto.getEmail(), encode);
+    // Save Repository
     User saved = userRepository.save(user);
     return SignUpResponseDto.from(saved);
   }
@@ -43,12 +50,19 @@ public class UserServiceImpl implements UserService {
   @Transactional
   @Override
   public void update(Long id, UpdateRequestDto dto) {
-    // id 를 먼저 찾기 , 단, id 와 pwd 동시에 찾기.
-    User findUser = userRepository.findByIdAndPwdOrElseThrow(id,dto.getPwd());
-
-    // 그 후 update 해주기
+    // dto valid 확인
     if (!dto.isValid()) {
       throw new InvalidRequestException("입력값이 잘못되었습니다.");
+    }
+
+    // id 에 해당하는 유저 조회
+    User findUser = userRepository.findByIdOrElseThrow(id);
+
+    // 해당 유저의 pwd 를 비교하기
+    Boolean isMatch = passwordEncoder.matches(dto.getPwd(), findUser.getPwd());
+
+    if(!isMatch) {
+      throw new InvalidPasswordException("비밀번호가 틀렸습니다.");
     }
 
     // 만약 널 값 ?비어있다면, 그냥 기존의 값을 유지하도록
@@ -62,14 +76,31 @@ public class UserServiceImpl implements UserService {
 
   @Override
   public void delete(Long id, DeleteRequestDto dto) {
-    User findUser = userRepository.findByIdAndPwdOrElseThrow(id, dto.getPwd());
+    // id 에 해당하는 유저 조회
+    User findUser = userRepository.findByIdOrElseThrow(id);
+
+    // 해당 유저의 pwd 를 비교하기
+    Boolean isMatch = passwordEncoder.matches(dto.getPwd(), findUser.getPwd());
+
+    if(!isMatch) {
+      throw new InvalidPasswordException("비밀번호가 틀렸습니다.");
+    }
+
     userRepository.delete(findUser);
   }
 
+  @Transactional
   @Override
   public FindResponseDto login(LoginRequestDto dto) {
-    // 아이디랑, 찾은 후에 ? return 해주면 됩니다.
-    User findUser = userRepository.findByEmailAndPwdOrElseThrow(dto.getEmail(), dto.getPwd());
+    // 이메일에 맞는 유저 찾기
+    User findUser = userRepository.findByEmailOrElseThrow(dto.getEmail());
+
+    // 유저의 비밀번호를 가져와서 비교
+    boolean isMatch = passwordEncoder.matches(dto.getPwd(), findUser.getPwd());
+
+    if(!isMatch) {
+      throw new InvalidPasswordException("비밀번호가 틀렸습니다.");
+    }
     return FindResponseDto.from(findUser);
   }
 }
